Refactors corrupt_value.py into a package
Refactors corrupt_value.py into a package and splits each Class into
it's own module
