+++ src/main/python/crptr/corrupt_values/corrupt_missing_value.py
@@ -1,0 +1,57 @@
+from crptr import base_functions
+from crptr.corrupt_values.base import CorruptValue
+
+
+class CorruptMissingValue(CorruptValue):
+  """A corruptor method which simply sets an attribute value to a missing
+     value.
+
+     The additional argument (besides the base class argument
+     'position_function') that has to be set when this attribute type is
+     initialised are:
+
+     missing_val  The string which designates a missing value. Default value
+                  is the empty string ''.
+
+     Note that the 'position_function' is not required by this corruptor
+     method.
+  """
+
+  # ---------------------------------------------------------------------------
+
+  def __init__(self, **kwargs):
+    """Constructor. Process the derived keywords first, then call the base
+       class constructor.
+    """
+
+    self.missing_val = ''
+    self.name =        'Missing value'
+
+    def dummy_position(s):  # Define a dummy position function
+      return 0
+
+    # Process all keyword arguments
+    #
+    base_kwargs = {}  # Dictionary, will contain unprocessed arguments
+
+    for (keyword, value) in list(kwargs.items()):
+
+      if (keyword.startswith('miss')):
+        base_functions.check_is_string('missing_val', value)
+        self.missing_val = value
+
+      else:
+        base_kwargs[keyword] = value
+
+    base_kwargs['position_function'] = dummy_position
+
+    CorruptValue.__init__(self, base_kwargs)  # Process base arguments
+
+  # ---------------------------------------------------------------------------
+
+  def corrupt_value(self, in_str):
+    """Simply return the missing value string.
+    """
+
+    return self.missing_val
+
